SparkyFitnessWhoop/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.12.5
python-dotenv==1.0.0
requests==2.32.5
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1